Make os-path-split always operate Unix-like
It seems that within LilyPond all paths have already been converted
to use forward slashes, so the differentiation in os-path-split is
not necessary. It will now always split on forward slashes.

A function os-path-split-os is in place as a fallback solution if it should
turn out to be necessary in some place or another.